Move PDSI calculation script here
For a handful of sites, the ecosystem runs a constant water deficit for
a couple hundred years. This causes CRAZY PDSI values if we use the
standard calibration period of 1931-1990, so I ran the values with a
couple different calibration periods.  The patterns of pdsi are
generally the same across calibration types, but the values are most
reasonably if we calibrate to the 1800-1850 time period.