Look at problem of uneven temporal extent of LBDA
The Living Blended Drought Atlas only goes back ~400 years in much of
the upper midwest.  The GCM output had very similar patterns of climate
stability in the past 400 years and past 1000, or at least patterns
that were consistent when we matched extents to the LBDA. Turns out
that when we use climate change over the past 400 years as a proxy for
the past 1000, there’s not a great correlation for much of the region
(R2=0.51; 0.61 for Upper Midwest).  This means we need to redo a lot of
work with the empirical data to make sure we’re matched in time during
our analyses.